The main concept of the table class was changed.
There is a new concept that consist of the idea of changing sides
automatically. The user shouldn't care about it anymore.
The user has specify the concrete Mode of a table and then just add new
tables. If the user has to add a table of a different mode they'll
have to change only Mode and keep adding tables. The object has to
determine if it's appropriate time to change the side or keep using
the current one.